72: zip file import doesnt work anymore
Task-Url: https://github.com/Polymap4/polymap4-p4/issues/issue/72

Have to handle case, that there are more than one file in context-in,
when delegating from archive importer to importers of the unpacked
files.

Import-Entry is removed from list (so still the problem with list
update, that there are remaining empty entries in list causing
exceptions clicking on them) when import was successful and the start
panel is only shown, when there are no more other files to import on
this site.